bf2stats (mysql) support, vip rank support (server accepts rank - python rank initialization still missing)

+++ emulator/playerdb.cpp
@@ -17,16 +17,16 @@ PlayerData::PlayerData(const std::string_view& name, const std::string_view& ema
 
 }
 
-PlayerData::PlayerData(std::uint32_t id, const std::string_view& name, const std::string_view& email, const std::string_view& password, const std::string_view& country)
+PlayerData::PlayerData(std::uint64_t id, const std::string_view& name, const std::string_view& email, const std::string_view& password, const std::string_view& country)
 	: id(id), name(name), email(email), password(password), country(country)
 {
 
 }
 
-std::uint32_t PlayerData::GetUserID() const {
+std::uint64_t PlayerData::GetUserID() const {
 	return id;
 }
 
-std::uint32_t PlayerData::GetProfileID() const {
+std::uint64_t PlayerData::GetProfileID() const {
 	return id;
 }
